feat: separate the upload-specific routes and update the respective sdk methods using it.
this PR closes #119
