+++ server/src/index.ts
@@ -10,6 +10,7 @@ import { adminRouter } from "./routes/admin.route.js";
 import { jobs as jobsRouter } from "./routes/jobs.route.js";
 import { serverRouter } from "./routes/server.route.js";
 import { solanaRouter } from "./routes/solana.route.js";
+import { uploadsRouter } from "./routes/upload.route.js";
 import { userRouter } from "./routes/user.route.js";
 import { ensureConfigInitialized } from "./utils/solana/index.js";
 
@@ -44,6 +45,7 @@ app.use(express.json());
 app.use(apiLimiter);
 
 app.use("/api/admin", adminRouter);
+app.use("/api/upload", uploadsRouter);
 app.use("/api/user", userRouter);
 app.use("/api/solana", solanaRouter);
 app.use("/api/jobs", jobsRouter);
